+++ docs/DryIoc.Docs/VersionHistory.md
@@ -1,10 +1,17 @@
 Version History
 ---------------
 
-## v5.1.0 Small feature and bug-fix release / 2022-05-20
+## v5.2.0 Small feature and bug-fix release / 2022-08-02
 
-- fixed: #486 RegisterInstance in principle does not support non Singleton reuse but InstanceFactory still allows to pass different reuse
-- added: #487 Create CreateChild overload exposing RegistrySharing and childKey arguments
+### Features
+
+- Adding `WithConcreteTypeDynamicRegistrations` overload with `IfUnresolved ifConcreteTypeIsUnresolved` parameter 
+to control exception information, and rule fallback behavior (#506)
+- Small speed-ups 
+
+### Fixes
+
+- #507 Transient resolve with opening scope using factory func in singleton
 
 
 ## v5.0.2 Bug-fix release / 2022-05-10